% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datafiles_functions.R
\name{datafile_content}
\alias{datafile_content}
\title{datafiles_content}
\usage{
datafile_content(api_key, datafile_id)
}
\arguments{
\item{api_key}{Your Datashift api_key}

\item{datafile_id}{The id of the file to be retrieved.}
}
\description{
Retrieves the content of a datafile from datashift.ca
}
\examples{
datafiles_content(api_key="123abc", datafile_id="12345")
}
\keyword{content}
\keyword{datafile}
\keyword{datashift}

