% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datafiles_functions.R
\name{datafiles_list}
\alias{datafiles_list}
\title{datafiles_list}
\usage{
datafiles_list(api_key)
}
\arguments{
\item{api_key}{Your Datashift api_key}
}
\description{
Retrieves a list of your datafiles from datashift.ca.
}
\examples{
datafiles_list(api_key="123abc")
}
\keyword{datafile}
\keyword{datashift}
\keyword{list}

